update: label data projections
We use the actual raw data as labels, limiting the size of the labels to
some rather randomly selected value. We truncate the input string to the
number of max characters.

If the label is provided in the POST request we prepend it to the final
label and truncate to the max char.

Signed-off-by: Milos Gajdos <milosthegajdos@gmail.com>